feat(api): add ElementForVerificationFactory
