+++ api/tests/devcomp/unit/domain/models/module/Module_test.js
@@ -1,11 +1,5 @@
-import { ModuleInstantiationError } from '../../../../../../src/devcomp/domain/errors.js';
-import { QCMForAnswerVerification } from '../../../../../../src/devcomp/domain/models/element/QCM-for-answer-verification.js';
-import { QCUForAnswerVerification } from '../../../../../../src/devcomp/domain/models/element/QCU-for-answer-verification.js';
-import { QROCMForAnswerVerification } from '../../../../../../src/devcomp/domain/models/element/QROCM-for-answer-verification.js';
 import { Module } from '../../../../../../src/devcomp/domain/models/module/Module.js';
-import { NotFoundError } from '../../../../../../src/shared/domain/errors.js';
-import { logger } from '../../../../../../src/shared/infrastructure/utils/logger.js';
-import { catchErrSync, expect, sinon } from '../../../../../test-helper.js';
+import { expect } from '../../../../../test-helper.js';
 
 describe('Unit | Devcomp | Domain | Models | Module | Module', function () {
   describe('#constructor', function () {
@@ -86,683 +80,6 @@ describe('Unit | Devcomp | Domain | Models | Module | Module', function () {
               grains: [],
             }),
         ).to.throw('The details are required for a module');
-      });
-    });
-
-    describe('#getGrainByElementId', function () {
-      it('should return the parent grain of the element with given id', function () {
-        // given
-        const elementId = 'elementId';
-        const id = 1;
-        const slug = 'les-adresses-email';
-        const title = 'Les adresses email';
-        const component = { type: 'element', element: { id: elementId } };
-        const expectedGrain = {
-          components: [component],
-        };
-        const details = Symbol('details');
-
-        // when
-        const foundGrain = new Module({
-          id,
-          slug,
-          title,
-          grains: [expectedGrain],
-          details,
-        }).getGrainByElementId(elementId);
-
-        // then
-        expect(foundGrain).to.deep.equal(expectedGrain);
-      });
-
-      it('should throw an error if grain does not exist', function () {
-        // given
-        const elementId = 'elementId';
-        const id = 1;
-        const slug = 'les-adresses-email';
-        const title = 'Les adresses email';
-        const component = { type: 'element', element: { id: elementId } };
-        const grain = {
-          components: [component],
-        };
-        const details = Symbol('details');
-        const module = new Module({ id, slug, title, grains: [grain], details });
-
-        // when
-        const error = catchErrSync(module.getGrainByElementId, module)('another-grain-id');
-
-        // then
-        expect(error).to.be.instanceOf(NotFoundError);
-      });
-    });
-  });
-
-  describe('#toDomainForVerification', function () {
-    it('should throw an ModuleInstantiationError if data is incorrect', function () {
-      // given
-      const feedbacks = { valid: 'valid', invalid: 'invalid' };
-      const proposals = [
-        { id: '1', content: 'toto' },
-        { id: '2', content: 'foo' },
-      ];
-
-      const dataWithMissingSolutionForQCU = {
-        id: '6282925d-4775-4bca-b513-4c3009ec5886',
-        slug: 'title',
-        title: 'title',
-        details: {
-          image: 'https://images.pix.fr/modulix/placeholder-details.svg',
-          description: 'Description',
-          duration: 5,
-          level: 'Débutant',
-          objectives: ['Objective 1'],
-        },
-        grains: [
-          {
-            id: 'f312c33d-e7c9-4a69-9ba0-913957b8f7dd',
-            type: 'lesson',
-            title: 'title',
-            components: [
-              {
-                type: 'element',
-                element: {
-                  id: '123',
-                  instruction: 'instruction',
-                  locales: ['fr-FR'],
-                  proposals,
-                  feedbacks,
-                  type: 'qcu',
-                },
-              },
-            ],
-          },
-        ],
-      };
-
-      // when
-      const error = catchErrSync(Module.toDomainForVerification)(dataWithMissingSolutionForQCU);
-
-      // then
-      expect(error).to.be.an.instanceOf(ModuleInstantiationError);
-      expect(error.message).to.deep.equal('The solution is required for a verification QCU');
-    });
-
-    describe('with answerable elements', function () {
-      describe('with ComponentElement', function () {
-        it('should instantiate a Module with components', function () {
-          // given
-          const feedbacks = { valid: 'valid', invalid: 'invalid' };
-          const proposals = [
-            { id: '1', content: 'toto' },
-            { id: '2', content: 'foo' },
-          ];
-
-          const moduleData = {
-            id: '6282925d-4775-4bca-b513-4c3009ec5886',
-            slug: 'title',
-            title: 'title',
-            details: {
-              image: 'https://images.pix.fr/modulix/placeholder-details.svg',
-              description: 'Description',
-              duration: 5,
-              level: 'Débutant',
-              objectives: ['Objective 1'],
-            },
-            grains: [
-              {
-                id: 'f312c33d-e7c9-4a69-9ba0-913957b8f7dd',
-                type: 'lesson',
-                title: 'title',
-                components: [
-                  {
-                    type: 'element',
-                    element: {
-                      id: '123',
-                      instruction: 'instruction',
-                      locales: ['fr-FR'],
-                      proposals,
-                      feedbacks,
-                      type: 'qcu',
-                      solution: proposals[0].id,
-                    },
-                  },
-                ],
-              },
-            ],
-          };
-
-          // when
-          const module = Module.toDomainForVerification(moduleData);
-
-          // then
-          expect(module).to.be.an.instanceOf(Module);
-          expect(module.grains).not.to.be.empty;
-          for (const grain of module.grains) {
-            expect(grain.components).not.to.be.empty;
-          }
-        });
-
-        it('should instantiate a Module with a ComponentElement with contains a QCUForAnswerVerification Element', function () {
-          // given
-          const feedbacks = { valid: 'valid', invalid: 'invalid' };
-          const proposals = [
-            { id: '1', content: 'toto' },
-            { id: '2', content: 'foo' },
-          ];
-
-          const moduleData = {
-            id: '6282925d-4775-4bca-b513-4c3009ec5886',
-            slug: 'title',
-            title: 'title',
-            details: {
-              image: 'https://images.pix.fr/modulix/placeholder-details.svg',
-              description: 'Description',
-              duration: 5,
-              level: 'Débutant',
-              objectives: ['Objective 1'],
-            },
-            grains: [
-              {
-                id: 'f312c33d-e7c9-4a69-9ba0-913957b8f7dd',
-                type: 'lesson',
-                title: 'title',
-                components: [
-                  {
-                    type: 'element',
-                    element: {
-                      id: '123',
-                      instruction: 'instruction',
-                      locales: ['fr-FR'],
-                      proposals,
-                      feedbacks,
-                      type: 'qcu',
-                      solution: proposals[0].id,
-                    },
-                  },
-                ],
-              },
-            ],
-          };
-
-          // when
-          const module = Module.toDomainForVerification(moduleData);
-
-          // then
-          expect(module.grains[0].components[0].element).to.be.an.instanceOf(QCUForAnswerVerification);
-        });
-
-        it('should instantiate a Module with a ComponentElement with contains a QCMForAnswerVerification Element', function () {
-          // given
-          const moduleData = {
-            id: '6282925d-4775-4bca-b513-4c3009ec5886',
-            slug: 'title',
-            title: 'title',
-            details: {
-              image: 'https://images.pix.fr/modulix/placeholder-details.svg',
-              description: 'Description',
-              duration: 5,
-              level: 'Débutant',
-              objectives: ['Objective 1'],
-            },
-            grains: [
-              {
-                id: 'f312c33d-e7c9-4a69-9ba0-913957b8f7dd',
-                type: 'lesson',
-                title: 'title',
-                components: [
-                  {
-                    type: 'element',
-                    element: {
-                      id: '30701e93-1b4d-4da4-b018-fa756c07d53f',
-                      type: 'qcm',
-                      instruction: '<p>Quels sont les 3 piliers de Pix ?</p>',
-                      proposals: [
-                        {
-                          id: '1',
-                          content: 'Evaluer ses connaissances et savoir-faire sur 16 compétences du numérique',
-                        },
-                        {
-                          id: '2',
-                          content: 'Développer son savoir-faire sur les jeux de type TPS',
-                        },
-                        {
-                          id: '3',
-                          content: 'Développer ses compétences numériques',
-                        },
-                        {
-                          id: '4',
-                          content: 'Certifier ses compétences Pix',
-                        },
-                        {
-                          id: '5',
-                          content: 'Evaluer ses compétences de logique et compréhension mathématique',
-                        },
-                      ],
-                      feedbacks: {
-                        valid: '<p>Correct ! Vous nous avez bien cernés :)</p>',
-                        invalid: '<p>Et non ! Pix sert à évaluer, certifier et développer ses compétences numériques.',
-                      },
-                      solutions: ['1', '3', '4'],
-                    },
-                  },
-                ],
-              },
-            ],
-          };
-
-          // when
-          const module = Module.toDomainForVerification(moduleData);
-
-          // then
-          expect(module.grains[0].components[0].element).to.be.an.instanceOf(QCMForAnswerVerification);
-        });
-
-        it('should instantiate a Module with a ComponentElement with contains a QROCMForAnswerVerification Element', function () {
-          // given
-          const moduleData = {
-            id: '6282925d-4775-4bca-b513-4c3009ec5886',
-            slug: 'title',
-            title: 'title',
-            details: {
-              image: 'https://images.pix.fr/modulix/placeholder-details.svg',
-              description: 'Description',
-              duration: 5,
-              level: 'Débutant',
-              objectives: ['Objective 1'],
-            },
-            grains: [
-              {
-                id: 'f312c33d-e7c9-4a69-9ba0-913957b8f7dd',
-                type: 'lesson',
-                title: 'title',
-                components: [
-                  {
-                    type: 'element',
-                    element: {
-                      id: '98c51fa7-03b7-49b1-8c5e-49341d35909c',
-                      type: 'qrocm',
-                      instruction:
-                        "<p>Pour être sûr que tout est clair, complétez le texte ci-dessous <span aria-hidden='true'>🧩</span></p><p>Si vous avez besoin d’aide, revenez en arrière <span aria-hidden='true'>⬆️</span></p>",
-                      proposals: [
-                        {
-                          type: 'text',
-                          content: '<p>Le symbole</>',
-                        },
-                        {
-                          input: 'symbole',
-                          type: 'input',
-                          inputType: 'text',
-                          size: 1,
-                          display: 'inline',
-                          placeholder: '',
-                          ariaLabel: 'Réponse 1',
-                          defaultValue: '',
-                          tolerances: ['t1'],
-                          solutions: ['@'],
-                        },
-                        {
-                          input: 'premiere-partie',
-                          type: 'select',
-                          display: 'inline',
-                          placeholder: '',
-                          ariaLabel: 'Réponse 2',
-                          defaultValue: '',
-                          tolerances: [],
-                          options: [
-                            {
-                              id: '1',
-                              content: "l'identifiant",
-                            },
-                            {
-                              id: '2',
-                              content: "le fournisseur d'adresse mail",
-                            },
-                          ],
-                          solutions: ['1'],
-                        },
-                      ],
-                      feedbacks: {
-                        valid: '<p>Bravo ! 🎉 </p>',
-                        invalid: "<p class='pix-list-inline'>Et non !</p>",
-                      },
-                    },
-                  },
-                ],
-              },
-            ],
-          };
-
-          // when
-          const module = Module.toDomainForVerification(moduleData);
-
-          // then
-          expect(module.grains[0].components[0].element).to.be.an.instanceOf(QROCMForAnswerVerification);
-        });
-      });
-
-      describe('with ComponentStepper', function () {
-        it('should instantiate a Module with components', function () {
-          // given
-          const feedbacks = { valid: 'valid', invalid: 'invalid' };
-          const proposals = [
-            { id: '1', content: 'toto' },
-            { id: '2', content: 'foo' },
-          ];
-
-          const moduleData = {
-            id: '6282925d-4775-4bca-b513-4c3009ec5886',
-            slug: 'title',
-            title: 'title',
-            details: {
-              image: 'https://images.pix.fr/modulix/placeholder-details.svg',
-              description: 'Description',
-              duration: 5,
-              level: 'Débutant',
-              objectives: ['Objective 1'],
-            },
-            grains: [
-              {
-                id: 'f312c33d-e7c9-4a69-9ba0-913957b8f7dd',
-                type: 'lesson',
-                title: 'title',
-                components: [
-                  {
-                    type: 'stepper',
-                    steps: [
-                      {
-                        elements: [
-                          {
-                            id: '123',
-                            instruction: 'instruction',
-                            locales: ['fr-FR'],
-                            proposals,
-                            feedbacks,
-                            type: 'qcu',
-                            solution: proposals[0].id,
-                          },
-                        ],
-                      },
-                    ],
-                  },
-                ],
-              },
-            ],
-          };
-
-          // when
-          const module = Module.toDomainForVerification(moduleData);
-
-          // then
-          expect(module).to.be.an.instanceOf(Module);
-          expect(module.grains).not.to.be.empty;
-          for (const grain of module.grains) {
-            expect(grain.components).not.to.be.empty;
-          }
-        });
-
-        it('should instantiate a Module with a ComponentStepper with contains a QCUForAnswerVerification Element', function () {
-          // given
-          const feedbacks = { valid: 'valid', invalid: 'invalid' };
-          const proposals = [
-            { id: '1', content: 'toto' },
-            { id: '2', content: 'foo' },
-          ];
-
-          const moduleData = {
-            id: '6282925d-4775-4bca-b513-4c3009ec5886',
-            slug: 'title',
-            title: 'title',
-            details: {
-              image: 'https://images.pix.fr/modulix/placeholder-details.svg',
-              description: 'Description',
-              duration: 5,
-              level: 'Débutant',
-              objectives: ['Objective 1'],
-            },
-            grains: [
-              {
-                id: 'f312c33d-e7c9-4a69-9ba0-913957b8f7dd',
-                type: 'lesson',
-                title: 'title',
-                components: [
-                  {
-                    type: 'stepper',
-                    steps: [
-                      {
-                        elements: [
-                          {
-                            id: '123',
-                            instruction: 'instruction',
-                            locales: ['fr-FR'],
-                            proposals,
-                            feedbacks,
-                            type: 'qcu',
-                            solution: proposals[0].id,
-                          },
-                        ],
-                      },
-                    ],
-                  },
-                ],
-              },
-            ],
-          };
-
-          // when
-          const module = Module.toDomainForVerification(moduleData);
-
-          // then
-          expect(module.grains[0].components[0].element).to.be.an.instanceOf(QCUForAnswerVerification);
-        });
-
-        it('should instantiate a Module with a ComponentStepper with contains a QCMForAnswerVerification Element', function () {
-          // given
-          const moduleData = {
-            id: '6282925d-4775-4bca-b513-4c3009ec5886',
-            slug: 'title',
-            title: 'title',
-            details: {
-              image: 'https://images.pix.fr/modulix/placeholder-details.svg',
-              description: 'Description',
-              duration: 5,
-              level: 'Débutant',
-              objectives: ['Objective 1'],
-            },
-            grains: [
-              {
-                id: 'f312c33d-e7c9-4a69-9ba0-913957b8f7dd',
-                type: 'lesson',
-                title: 'title',
-                components: [
-                  {
-                    type: 'stepper',
-                    steps: [
-                      {
-                        elements: [
-                          {
-                            id: '30701e93-1b4d-4da4-b018-fa756c07d53f',
-                            type: 'qcm',
-                            instruction: '<p>Quels sont les 3 piliers de Pix ?</p>',
-                            proposals: [
-                              {
-                                id: '1',
-                                content: 'Evaluer ses connaissances et savoir-faire sur 16 compétences du numérique',
-                              },
-                              {
-                                id: '2',
-                                content: 'Développer son savoir-faire sur les jeux de type TPS',
-                              },
-                              {
-                                id: '3',
-                                content: 'Développer ses compétences numériques',
-                              },
-                              {
-                                id: '4',
-                                content: 'Certifier ses compétences Pix',
-                              },
-                              {
-                                id: '5',
-                                content: 'Evaluer ses compétences de logique et compréhension mathématique',
-                              },
-                            ],
-                            feedbacks: {
-                              valid: '<p>Correct ! Vous nous avez bien cernés :)</p>',
-                              invalid:
-                                '<p>Et non ! Pix sert à évaluer, certifier et développer ses compétences numériques.',
-                            },
-                            solutions: ['1', '3', '4'],
-                          },
-                        ],
-                      },
-                    ],
-                  },
-                ],
-              },
-            ],
-          };
-
-          // when
-          const module = Module.toDomainForVerification(moduleData);
-
-          // then
-          expect(module.grains[0].components[0].element).to.be.an.instanceOf(QCMForAnswerVerification);
-        });
-
-        it('should instantiate a Module with a ComponentStepper with contains a QROCMForAnswerVerification Element', function () {
-          // given
-          const moduleData = {
-            id: '6282925d-4775-4bca-b513-4c3009ec5886',
-            slug: 'title',
-            title: 'title',
-            details: {
-              image: 'https://images.pix.fr/modulix/placeholder-details.svg',
-              description: 'Description',
-              duration: 5,
-              level: 'Débutant',
-              objectives: ['Objective 1'],
-            },
-            grains: [
-              {
-                id: 'f312c33d-e7c9-4a69-9ba0-913957b8f7dd',
-                type: 'lesson',
-                title: 'title',
-                components: [
-                  {
-                    type: 'stepper',
-                    steps: [
-                      {
-                        elements: [
-                          {
-                            id: '98c51fa7-03b7-49b1-8c5e-49341d35909c',
-                            type: 'qrocm',
-                            instruction:
-                              "<p>Pour être sûr que tout est clair, complétez le texte ci-dessous <span aria-hidden='true'>🧩</span></p><p>Si vous avez besoin d’aide, revenez en arrière <span aria-hidden='true'>⬆️</span></p>",
-                            proposals: [
-                              {
-                                type: 'text',
-                                content: '<p>Le symbole</>',
-                              },
-                              {
-                                input: 'symbole',
-                                type: 'input',
-                                inputType: 'text',
-                                size: 1,
-                                display: 'inline',
-                                placeholder: '',
-                                ariaLabel: 'Réponse 1',
-                                defaultValue: '',
-                                tolerances: ['t1'],
-                                solutions: ['@'],
-                              },
-                              {
-                                input: 'premiere-partie',
-                                type: 'select',
-                                display: 'inline',
-                                placeholder: '',
-                                ariaLabel: 'Réponse 2',
-                                defaultValue: '',
-                                tolerances: [],
-                                options: [
-                                  {
-                                    id: '1',
-                                    content: "l'identifiant",
-                                  },
-                                  {
-                                    id: '2',
-                                    content: "le fournisseur d'adresse mail",
-                                  },
-                                ],
-                                solutions: ['1'],
-                              },
-                            ],
-                            feedbacks: {
-                              valid: '<p>Bravo ! 🎉 </p>',
-                              invalid: "<p class='pix-list-inline'>Et non !</p>",
-                            },
-                          },
-                        ],
-                      },
-                    ],
-                  },
-                ],
-              },
-            ],
-          };
-
-          // when
-          const module = Module.toDomainForVerification(moduleData);
-
-          // then
-          expect(module.grains[0].components[0].element).to.be.an.instanceOf(QROCMForAnswerVerification);
-        });
-      });
-    });
-
-    describe('without answerable elements', function () {
-      it('should instantiate a Module with no ComponentElement and log warning', function () {
-        // given
-        const moduleData = {
-          id: '6282925d-4775-4bca-b513-4c3009ec5886',
-          slug: 'title',
-          title: 'title',
-          details: {
-            image: 'https://images.pix.fr/modulix/placeholder-details.svg',
-            description: 'Description',
-            duration: 5,
-            level: 'Débutant',
-            objectives: ['Objective 1'],
-          },
-          grains: [
-            {
-              id: 'f312c33d-e7c9-4a69-9ba0-913957b8f7dd',
-              type: 'lesson',
-              title: 'title',
-              components: [
-                {
-                  type: 'element',
-                  element: {
-                    id: '84726001-1665-457d-8f13-4a74dc4768ea',
-                    type: 'text',
-                    content: '<h3>Content</h3>',
-                  },
-                },
-              ],
-            },
-          ],
-        };
-        sinon.stub(logger, 'warn').returns();
-
-        // when
-        const module = Module.toDomainForVerification(moduleData);
-
-        // then
-        expect(logger.warn).to.have.been.calledWithExactly({
-          event: 'module_element_type_not_handled_for_verification',
-          message: `Element type not handled for verification: ${moduleData.grains[0].components[0].element.type}`,
-        });
-        expect(module).to.be.an.instanceOf(Module);
-        expect(module.grains).not.to.be.empty;
-        expect(module.grains[0].components).to.deep.equal([]);
       });
     });
   });
